requirements.txt
bs4==0.0.1
celery[redis]==5.2.7
Django==4.1.3
djangorestframework==3.14.0
gunicorn==20.1.0
loguru==0.6.0
newspaper3k==0.2.8
psycopg2-binary==2.9.5
pybrake==1.10.0
requests==2.28.1
sentry-sdk==1.11.0
six==1.16.0
whitenoise==6.2.0
